+++ packages/sdks/core-js-sdk/src/httpClient/index.ts
@@ -1,4 +1,4 @@
-import { transformSetCookie } from './helpers';
+import { getClientSessionId, transformSetCookie } from './helpers';
 import createFetchLogger from './helpers/createFetchLogger';
 import {
   CreateHttpClientConfig,
@@ -11,34 +11,6 @@ import { mergeHeaders, serializeBody } from './utils';
 
 const jsonHeaders = {
   'Content-Type': 'application/json',
-};
-
-let sessionId: string;
-const getClientSessionId = (): string => {
-  if (sessionId) {
-    return sessionId;
-  }
-  const currentDate = new Date();
-  const utcString = `${currentDate.getUTCFullYear().toString()}-${(
-    currentDate.getUTCMonth() + 1
-  )
-    .toString()
-    .padStart(2, '0')}-${currentDate
-    .getUTCDate()
-    .toString()
-    .padStart(2, '0')}-${currentDate
-    .getUTCHours()
-    .toString()
-    .padStart(2, '0')}:${currentDate
-    .getUTCMinutes()
-    .toString()
-    .padStart(2, '0')}:${currentDate
-    .getUTCSeconds()
-    .toString()
-    .padStart(2, '0')}:${currentDate.getUTCMilliseconds().toString()}`;
-  const randomSuffix = Math.floor(1000 + Math.random() * 9000);
-  sessionId = `${utcString}-${randomSuffix}`;
-  return sessionId;
 };
 
 /**
